+++ transtats/dashboard/managers/settings.py
@@ -23,26 +23,15 @@ from ..models.transplatform import TransPlatform
 from .base import BaseManager
 
 
-class TransPlatformManager(BaseManager):
+class AppSettingsManager(BaseManager):
     """
-    Translation Platform Manager
+    Application Settings Manager
     """
 
-    def get_translation_platform(self):
+    def get_translation_platforms(self):
         """
         fetch translation platforms from db, zanata as of now!
         """
-        platform = {}
-        '''
-        zanata_platform = TransPlatform(engine_name='zanata', api_url='http://translate.zanata.org',
-                                        server_status='active', subject='public', platform_id=3)
-        self.db_session.add(zanata_platform)
-        self.db_session.commit()
-        '''
-        translation_platform = self.db_session.query(TransPlatform).filter_by(subject='fedora').first()
-        if translation_platform:
-            platform['url'] = translation_platform.api_url
-            platform['engine'] = translation_platform.engine_name
-            platform['subject'] = translation_platform.subject
-            platform['state'] = translation_platform.server_status
-        return platform
+        platforms = {}
+        platform = self.db_session.query(TransPlatform).all()
+        return platforms