+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-##### Wed Jul 27 2016 Sundeep Anand <suanand@redhat.com>
+##### Fri Jul 29 2016 Sundeep Anand <suanand@redhat.com>
+* add alembic and settings page
 * rename project to transtats, fix ui and migrate to PostgreSQL
 * add features page, zanata api, view, manager and model
 * link travis and readmedocs